reafactor: handleCreate를 useSchedule로 이동
